% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces19w.R
\docType{data}
\name{ces19w}
\alias{ces19w}
\title{2019 Canadian Election Study Snippet}
\format{
An object of class \code{data.frame}
}
\usage{
data(ces19w)
}
\description{
Data from the 2019 Canadian Election Study Online Survey.
}
\references{
Stephenson, Laura B; Harell, Allison; Rubenson, Daniel; Loewen,
Peter John, 2020, "2019 Canadian Election Study - Online Survey",
https://doi.org/10.7910/DVN/DUS88V, Harvard Dataverse, V1.
}
\keyword{datasets}
