% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{low}
\alias{low}
\title{Lower 95\% CI quantile}
\usage{
low(x)
}
\arguments{
\item{x}{Value for which the 5th percentile should be returned.}
}
\value{
A numeric indicating the 2.5\% quantile
}
\description{
Lower 95\% CI quantile
}
