% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlSig.R
\name{mnlSig.svrepstatmisc}
\alias{mnlSig.svrepstatmisc}
\title{mnlSig method for \code{svrepstatmisc} objects}
\usage{
\method{mnlSig}{svrepstatmisc}(
  obj,
  pval = 0.05,
  two.sided = TRUE,
  flag.sig = TRUE,
  insig.blank = FALSE,
  ...
)
}
\arguments{
\item{obj}{A model object of class \code{svrepstatmisc}.}

\item{pval}{The desired Type I error rate to identify coefficients as
statistically significant.}

\item{two.sided}{Logical indicating whether calculated p-values should be
two-sided (if \code{TRUE}) or one-sided (if \code{FALSE}).}

\item{flag.sig}{Logical indicating whether an asterisk should be placed
beside coefficients which are significant at the \code{pval} level.}

\item{insig.blank}{Logical indicating whether coefficients which are not
significant at the \code{pval} level should be blank in the output.}

\item{...}{Other arguments (not currently implemented).}
}
\value{
A data frame suitable for printing with the (optionally
significance-flagged) coefficients from a multinomial logit model.
}
\description{
mnlSig method for \code{svrepstatmisc} objects
}
\author{
Dave Armstrong & John Santos
}
