% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.svyEffects}
\title{Plot method for \pkg{svyEffects} objects}
\usage{
\method{plot}{svyEffects}(x, what = c("preds", "diffs"), ...)
}
\arguments{
\item{x}{An object of class \pkg{svyEffects} generated by the either of
the functions \code{svyAME} or \code{svyMER}.}

\item{what}{A string denoting whether predictions or differences should be
plotted. These are the dataframes \code{$preds} and \code{$diffs}, which are
contained within the output generated by either \code{svyAME} or \code{svyMER}.}

\item{...}{Other arguments passed on the output.}
}
\value{
An object of class \code{ggplot}
}
\description{
Uses the \pkg{ggplot2} engine to visualize the predicted
probabilities or differences in predicted probabilities calculated by
\pkg{svyEffects}.
}
\examples{
data(ces19)
library(survey)
library(ggplot2)
ces19_svy <- svydesign(ids = ~1, strata = NULL, weights = ~pesweight,
  data = ces19, digits = 3)
VOTECON <- svyglm(votecon ~ agegrp + gender + educ + region + marketlib,
  design = ces19_svy, family = binomial)
ame_educ <- svyAME(VOTECON, varname = "educ", weightvar = "pesweight", seed = 2019)
plot(ame_educ)   # plots predicted probabilities
plot(ame_educ, "diffs")   # plots differences for pairwise comparisons

# Plots can be customized using ggplot's conventions, e.g.:
plot(ame_educ) +
  geom_hline(yintercept = 0, linetype = "dotted")
  labs(title = "mytitle",
       x = "x axis label",
       y = "y axis label",
       caption = "my caption") +
 theme_classic() +
 theme(axis.text.x = element_text(angle=45, hjust=1))

}
\author{
John Santos
}
