% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.svyEffects}
\title{Print method for \pkg{svyEffects} objects}
\usage{
\method{print}{svyEffects}(x, nrow = 40, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \pkg{svyEffects} generated by the either of
the functions \code{svyAME} or \code{svyMER}.}

\item{...}{Other arguments passed on the output.}
}
\value{
A dataframe
}
\description{
Prints results for \code{svyEffects} of objects
}
\examples{

}
\author{
John Santos
}
