% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_functions.R
\name{svyAME.multinom}
\alias{svyAME.multinom}
\title{Average marginal effects (a.k.a. marginal effects at observed values) for
(non-survey) weighted multinomial dependent-variable models (nnet::multinom).}
\usage{
\method{svyAME}{multinom}(
  obj,
  varname,
  weightvar,
  design,
  nvals = 11,
  diffchange = c("range", "unit", "sd"),
  sims = 1500,
  seed = NULL,
  ...
)
}
\arguments{
\item{obj}{Model object of class \code{svrepmisc::svrepstatmisc}.}

\item{varname}{Character string denoting the name of the predictor variable
for which effects are to be calculated.}

\item{weightvar}{A survey design object of class \code{survey::svydesign}.}

\item{design}{A survey design object with replicate weights of class
\code{survey::svyrep.design}.}

\item{nvals}{Scalar denoting the sequence length spanning the range of a
continuous variable for which effects are to be calculated (default: 11).}

\item{diffchange}{Character string  denoting over what change in x a first
difference is to be calculated for a continuous predictor (default: range).}

\item{sims}{Scalar denoting how many simulations to conduct (default: 1500).}

\item{seed}{Seed value for random number generator. By default, the function
picks a random integer between 1 and 1,000,000. If you save the output of
this function, it will save the seed value used for simulations in the slot
\code{$seed}.}

\item{...}{Other arguments (currently not implemented).}
}
\value{
A list with dataframes:
\describe{
\item{\code{$preds}}{predicted probabilities}
\item{\code{$preds}}{differences in predicted probabilities}
\item{\code{$seed}}{seed value used for random number generator}
}
}
\description{
Average marginal effects (a.k.a. marginal effects at observed values) for
(non-survey) weighted multinomial dependent-variable models (nnet::multinom).
}
