% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyAME.svyolr.R
\name{svyAME.svyolr}
\alias{svyAME.svyolr}
\title{Average Marginal Effects For Ordered Logit Models Of Survey-Weighted Data}
\usage{
\method{svyAME}{svyolr}(
  obj,
  varname,
  nvals = 11,
  diffchange = c("sd", "range", "unit"),
  byvar = NULL,
  bychange = c("sd", "range"),
  bynvals = 3,
  sims = 2500,
  seed = NULL,
  ci = 0.95,
  ...
)
}
\arguments{
\item{obj}{Model object of class \code{survey::svyolr}.}

\item{varname}{Character string denoting the name of the predictor variable
for which effects are to be calculated.}

\item{nvals}{Scalar denoting the sequence length spanning the range of a
continuous variable for which effects are to be calculated (default: 11).}

\item{diffchange}{Character string  denoting over what change in x a first
difference is to be calculated for a continuous predictor (default: "sd").}

\item{byvar}{For interaction effects, a character string denoting the name of
the moderator variable.}

\item{bychange}{For interaction effects with a numerical moderator, a
character string denoting for what values above and below the mean
predictions should be calculated (default: "sd").}

\item{bynvals}{For interaction effects with a numerical moderator, a scalar
denoting the sequence length  spanning the range of the moderator variable
for which effects are to be calculated (default: 3).}

\item{sims}{Scalar denoting how many simulations to conduct (default: 2500).}

\item{seed}{Seed value for random number generator. By default, the function
picks a random integer between 1 and 1,000,000. If you save the output of
this function, it will save the seed value used for simulations in the slot
\code{$seed}.}

\item{ci}{Scalar indicating confidence level to be used (default: .95).}

\item{...}{Other arguments (currently not implemented).}
}
\value{
A list with dataframes:
\describe{
\item{\code{$preds}}{predicted probabilities}
\item{\code{$diffs}}{differences in predicted probabilities}
\item{\code{$seed}}{seed value used for random number generator}
}
}
\description{
Calculates predicted probabilities and differences in probabilities
for a predictor holding all other observations at observed values
(i.e. a true average marginal effect). Uses simulations (the parametric
bootstrap) to derive 95\% confidence intervals.
}
\examples{
\dontrun{
data(ces19)
library(survey)
ces19_svy <- svydesign(ids = ~1, strata = NULL, weights = ~pesweight,
  data = ces19, digits = 3)
CONLDR <- svyolr(ftconldr ~ agegrp + gender + educ + region + marketlib,
  design = ces19_svy)
svyAME(CONLDR, varname = "region", seed = 2019)
svyAME(CONLDR, varname = "marketlib", seed = 2019)
}

}
\author{
John Santos & Dave Armstrong
}
