% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_functions.R
\name{svyMER}
\alias{svyMER}
\title{S3 generic function for survey-weighted marginal effects at reasonable values}
\usage{
svyMER(obj, ...)
}
\arguments{
\item{obj}{Model object on which to conduct post-estimation}

\item{...}{Other arguments (currently not implemented).}
}
\description{
S3 generic function for survey-weighted marginal effects at reasonable values
}
