% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_functions.R
\name{svyMER.svrepstatmisc}
\alias{svyMER.svrepstatmisc}
\title{Marginal effects at reasonable values for multinomial dependent-variable
models of survey-weighted data (svrepmisc::svrepstatmisc).}
\usage{
\method{svyMER}{svrepstatmisc}(
  obj,
  varname,
  weightvar,
  design,
  modform,
  nvals = 11,
  diffchange = c("range", "unit", "sd"),
  sims = 1500,
  seed = NULL,
  ...
)
}
\arguments{
\item{obj}{Model object of class \code{svrepmisc::svrepstatmisc}.}

\item{varname}{Character string denoting the name of the predictor variable
for which effects are to be calculated.}

\item{weightvar}{A survey design object of class \code{survey::svydesign}.}

\item{design}{A survey design object with replicate weights of class
\code{survey::svyrep.design}.}

\item{modform}{Character string denoting the model formula used of the model.
Must be in the format of class \code{modform = "y ~ x"} and match the model's
formula exactly.}

\item{nvals}{Scalar denoting the sequence length spanning the range of a
continuous variable for which effects are to be calculated (default: 11).}

\item{diffchange}{Character string  denoting over what change in x a first
difference is to be calculated for a continuous predictor (default: range).}

\item{sims}{Scalar denoting how many simulations to conduct (default: 1500).}

\item{seed}{Seed value for random number generator. By default, the function
picks a random integer between 1 and 1,000,000. If you save the output of
this function, it will save the seed value used for simulations in the slot
\code{$seed}.}

\item{...}{Other arguments (currently not implemented).}
}
\value{
A list with dataframes:
\describe{
\item{\code{$preds}}{predicted probabilities}
\item{\code{$preds}}{differences in predicted probabilities}
\item{\code{$seed}}{seed value used for random number generator}
\item{\code{$typical}}{the values at which variables are set for the simulations}
}
}
\description{
READ THIS FIRST: YOU NEED THE GITHUB PACKAGE \code{svrepmisc} TO USE THIS
FUNCTION!This function is designed to be used on a multinomial dependent
variable model of class \code{svrepmisc::svymultinom}) estimated on a survey
design object of class \code{survey::svrepstatmisc}). Currently, there are no
multinomial logit \code{R} functions that work on survey-weighted data.
However, Carl Ganz' \code{svrepmisc} package (available on GitHub,
https://github.com/carlganz/svrepmisc) uses replicate weights and jackknife
standard errors to approximate the same end result. (The output with this
approach is nearly identical to results obtained in Stata using either
sampling weights (i.e. "pweights") or replicate weights with jackknife
standard errors.)
}
\details{
To use this function, you use \code{survey::svydesign} to create a survey
design object with your sampling/design weights. Then, you create
survey design object with replicate weights. Then, you estimate your
multinomial logit model on the second survey design object using
\code{svrepmisc::svymultinom} (see the examples below).
}
