% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyMER.svyolr.R
\name{svyMER.svyolr}
\alias{svyMER.svyolr}
\title{Marginal effects at reasonable values for ordered logit models of survey-weighted data}
\usage{
\method{svyMER}{svyolr}(
  obj,
  varname,
  weightvar,
  nvals = 11,
  diffchange = c("range", "unit", "sd"),
  byvar = NULL,
  bynvals = 3,
  sims = 2500,
  seed = NULL,
  ...
)
}
\arguments{
\item{obj}{Model object of class \code{survey::svyolr}.}

\item{varname}{Character string denoting the name of the predictor variable
for which effects are to be calculated.}

\item{weightvar}{Character string denoting the name of the sampling weight  variable.}

\item{nvals}{Scalar denoting the sequence length spanning the range of a
continuous variable for which effects are to be calculated (default: 10).}

\item{diffchange}{Character string  denoting over what change in x a first
difference is to be calculated for a continuous predictor (default: range).}

\item{byvar}{For interaction effects, a character string denoting the name of
the moderator variable.}

\item{bynvals}{For interaction effects with a numerical moderator, a scalar
denoting the sequence length  spanning the range of the moderator variable
for which effects are to be calculated (default: 3).}

\item{sims}{Scalar denoting how many simulations to conduct (default: 2500).}

\item{seed}{Seed value for random number generator. By default, the function
picks a random integer between 1 and 1,000,000. If you save the output of
this function, it will save the seed value used for simulations in the slot
\code{$seed}.}

\item{...}{Other arguments (currently not implemented).}
}
\value{
A list with dataframes:
\describe{
\item{\code{$preds}}{predicted probabilities}
\item{\code{$diffs}}{differences in predicted probabilities}
\item{\code{$seed}}{seed value used for random number generator}
\item{\code{$typical}}{the values at which variables are set for the simulations}
}
}
\description{
Calculates predicted probabilities and differences in probabilities
for a predictor holding all other observations at reasonable/representative/typical
values (i.e. for an "average case"). This involves setting all continuous
variables to their medians and all categorical variables to their modes. Uses
simulations (the parametric bootstrap) to derive 95\% confidence intervals.
}
\examples{
data(ces19)
library(survey)
ces19_svy <- svydesign(ids = ~1, strata = NULL, weights = ~pesweight,
  data = ces19, digits = 3)
CONLDR <- svyolr(ftconldr ~ agegrp + gender + educ + region + marketlib,
  design = ces19_svy)
svyMER(CONLDR, varname = "region", weightvar = "pesweight", seed = 2019)
svyMER(CONLDR, varname = "marketlib", weightvar = "pesweight", seed = 2019)

}
\author{
John Santos & Dave Armstrong
}
