% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{svymode}
\alias{svymode}
\title{Find central/average value for all variables (for MERs)}
\usage{
svymode(x, design, ...)
}
\arguments{
\item{x}{Termlist.}

\item{design}{Survey design object.}

\item{...}{Other arguments (currently not implemented).}
}
\value{
A list for input into the svyMER functions.
}
\description{
Find central/average value for all variables (for MERs)
}
\author{
Dave Armstrong
}
