% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyPRE.R
\name{svyPRE.svrepstatmisc}
\alias{svyPRE.svrepstatmisc}
\title{Proportional reduction in error of survey-weighted multinomial logit models}
\usage{
\method{svyPRE}{svrepstatmisc}(obj, ...)
}
\arguments{
\item{obj}{Model object of class \code{svrepstatmisc}}

\item{...}{Other options (not currently implemented)}
}
\value{
A tibble
}
\description{
Proportional reduction in error of survey-weighted multinomial logit models
}
\examples{
\dontrun{
data(ces19)
library(survey)
ces19_svy <- svydesign(ids = ~1, strata = NULL, weights = ~pesweight,
  data = ces19, digits = 3)
ces19_svy_r <- as.svrepdesign(temp, type = "JK1")
# remotes::install_github("carlganz/svrepmisc")
library(svrepmisc)
VOTE <- svymultinom(vote ~ agegrp + gender + educ + region + relig + marketlib + culturetrad,
  design = ces19_svy_r, trace = FALSE)
svyPRE(VOTE)
}

}
\author{
John Santos & Dave Armstrong
}
