% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyPRE.R
\name{svyPRE.svyolr}
\alias{svyPRE.svyolr}
\title{Proportional reduction in error of survey-weighted ordered logit models}
\usage{
\method{svyPRE}{svyolr}(obj, ...)
}
\arguments{
\item{obj}{Model object of class \code{svyglm} with \code{family = "binomial"}
or \code{svyolr}}

\item{...}{Other options (not currently implemented)}
}
\value{
A tibble
}
\description{
Proportional reduction in error of survey-weighted ordered logit models
}
\examples{
data(ces19)
library(survey)
ces19_svy <- svydesign(ids = ~1, strata = NULL, weights = ~pesweight,
  data = ces19, digits = 3)
CONLDR <- svyolr(ftconldr ~ agegrp + gender + educ + region + marketlib,
  design = ces19_svy)
svyPRE(CONLDR)

}
\author{
John Santos & Dave Armstrong
}
