% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.svyEffects}
\title{Generic plot method for \pkg{svyEffects} objects}
\usage{
\method{plot}{svyEffects}(x, what = c("preds", "diffs"), ...)
}
\arguments{
\item{x}{An object of class \pkg{svyEffects} generated by the either of
the functions \code{svyAME} or \code{svyMER}.}

\item{what}{A string denoting whether predictions or differences should be
plotted. These are the dataframes \code{$preds} and \code{$diffs}, which are
contained within the output generated by either \code{svyAME} or \code{svyMER}.}

\item{...}{Other arguments passed on the output.}
}
\value{
An object of class \code{ggplot}
}
\description{
Uses the \pkg{ggplot2} engine to visualize the predicted
probabilities or differences in predicted probabilities calculated by
\pkg{svyEffects}.
}
