% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.svyEffects}
\title{Print method for \pkg{svyEffects} objects}
\usage{
\method{print}{svyEffects}(x, nrow = 40, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \pkg{svyEffects} generated by the either of
the functions \code{svyAME} or \code{svyMER}.}

\item{digits}{= Scalar indicating desired precision (in number digits to the right of the decimal point).}

\item{...}{Other arguments passed on the output.}
}
\value{
A list
}
\description{
Prints results for \code{svyEffects} of objects
}
\examples{
data(ces19)
library(survey)
library(ggplot2)
ces19_svy <- svydesign(ids = ~1, strata = NULL, weights = ~pesweight,
  data = ces19, digits = 3)
VOTECON <- svyglm(votecon ~ agegrp + gender + educ + region + marketlib,
  design = ces19_svy, family = binomial)
ame_educ <- svyAME(VOTECON, varname = "educ", weightvar = "pesweight", seed = 2019)
print(ame_educ)
}
\author{
John Santos
}
