% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{svyAME}
\alias{svyAME}
\title{S3 generic function for survey-weighted average marginal effects}
\usage{
svyAME(obj, ...)
}
\arguments{
\item{obj}{Model object on which to conduct post-estimation}

\item{...}{Other arguments (currently not implemented).}
}
\description{
S3 generic function for survey-weighted average marginal effects
}
