% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{svyMER}
\alias{svyMER}
\title{Marginal effects at reasonable values for binary, ordinal, and
multinomial dependent variable models of survey-weighted data}
\usage{
svyMER(obj, varname, ...)
}
\arguments{
\item{obj}{A model object. Currently, this package supports models of class
\code{svyglm} and \code{svyolr} from the \pkg{survey} package, and models of
class \code{svrepstatmisc} from the \pkg{svrepmisc} package (see note below).
Models of class \code{glm}, \code{polr} (from \pkg{MASS}), and \code{multinom}
(from \pkg{nnet}) that specify the \code{weight=} option also work.}

\item{varname}{Character string denoting the name of the predictor variable
for which effects are to be calculated.}

\item{...}{Other arguments used by svyMER methods.}
}
\description{
Calculates predicted probabilities for an independent variable,
holding all other observations at reasonable/representative/typical values
(i.e. for an "average case"). This involves setting all continuous variables
to their medians and all categorical variables to their modes. Uses
simulations (the parametric bootstrap) to derive 95\% confidence intervals.
Note: you will need to specify different arguments depending on whether your
model has a binary, ordinal, or categorical dependent variable, so consult
the documentation for this function's submethods for more details.
}
