% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyMER.svrepstatmisc.R
\name{svyMER.svrepstatmisc}
\alias{svyMER.svrepstatmisc}
\title{Marginal effects at reasonable values for multinomial logit models of survey-weighted data}
\usage{
\method{svyMER}{svrepstatmisc}(
  obj,
  varname,
  weightvar,
  nvals = 11,
  diffchange = c("range", "unit", "sd"),
  byvar = NULL,
  bynvals = 3,
  sims = 1500,
  seed = NULL,
  design,
  modform,
  ...
)
}
\arguments{
\item{obj}{Model object of class \code{svrepmisc::svrepstatmisc}.}

\item{varname}{Character string denoting the name of the predictor variable
for which effects are to be calculated.}

\item{weightvar}{A survey design object of class \code{survey::svydesign}.}

\item{nvals}{Scalar denoting the sequence length spanning the range of a
continuous variable for which effects are to be calculated (default: 11).}

\item{diffchange}{Character string  denoting over what change in x a first
difference is to be calculated for a continuous predictor (default: range).}

\item{byvar}{For interaction effects, a character string denoting the name of
the moderator variable.}

\item{bynvals}{For interaction effects with a numerical moderator, a scalar
denoting the sequence length  spanning the range of the moderator variable
for which effects are to be calculated (default: 3).}

\item{sims}{Scalar denoting how many simulations to conduct (default: 1500).}

\item{seed}{Seed value for random number generator. By default, the function
picks a random integer between 1 and 1,000,000. If you save the output of
this function, it will save the seed value used for simulations in the slot
\code{$seed}.}

\item{design}{A survey design object with replicate weights of class
\code{survey::svyrep.design}.}

\item{modform}{Character string denoting the model formula used of the model.
Must be in the format of class \code{modform = "y ~ x"} and match the model's
formula exactly.}

\item{...}{Other arguments (currently not implemented).}
}
\value{
A list with dataframes:
\describe{
\item{\code{$preds}}{predicted probabilities}
\item{\code{$diffs}}{differences in predicted probabilities}
\item{\code{$seed}}{seed value used for random number generator}
\item{\code{$typical}}{the values at which variables are set for the simulations}
}
}
\description{
Calculates predicted probabilities and differences in probabilities
for a predictor holding all other observations at reasonable/representative/typical
values (i.e. for an "average case"). This involves setting all continuous
variables to their medians and all categorical variables to their modes. Uses
simulations (the parametric bootstrap) to derive 95\% confidence intervals.

READ THIS BEFORE USING: You need the GitHub package \code{svrepmisc} to
use this function! Currently, there are no multinomial logit \code{R} functions
that work on survey-weighted data. However, Carl Ganz' \code{svrepmisc} package
(available here: \url{https://github.com/carlganz/svrepmisc}) uses replicate
weights and jackknife standard errors to approximate the same end result.
(The output with this approach is very similar to results obtained in Stata
using either sampling weights (i.e. "pweights") or replicate weights with
jackknife standard errors.)

To use this function...
\enumerate{
\item Use \code{survey::svydesign} to create a survey design object with your sampling/design weights.
\item Use \code{survey::as.svrepdesign(x, type="JK1")} convert the sampling weights to replicate weights with jackknife standard errors.
\item Estimate your multinomial logit model on the replicate weights design object using \code{svrepmisc::svymultinom}.
\item Now, you're ready to use \code{svyAME} to calculate probabilities.
}
}
\author{
John Santos & Dave Armstrong
}
